pyproject.toml
[tool.poetry]
name = "agent-conclave"
version = "0.1.0"
description = "GCP-native multi-agent conclave project"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
langgraph = "^0.0.26"
google-cloud-firestore = "^2.14.0"
google-cloud-secret-manager = "^2.18.0"
google-cloud-aiplatform = { extras = ["agent_engines", "adk"], version = ">=1.112.0" }
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.4"
typer = "^0.9.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
